+++ src/mg/itu/prom16/utils/ControllerUtils.java
@@ -1,0 +1,36 @@
+package mg.itu.prom16.utils;
+
+import java.io.File;
+import java.net.URLDecoder;
+import java.util.ArrayList;
+import java.util.List;
+
+public class ControllerUtils {
+    boolean isController(Class<?> c) {
+        return c.isAnnotationPresent(Controller.class);
+    }
+
+    public List<Class<?>> getAllControllers(String packageName) throws Exception {
+        List<Class<?>> res=new ArrayList<Class<?>>();
+        //répertoire racine du package
+        String path = this.getClass().getClassLoader().getResource(packageName.replace('.', '/')).getPath();
+        String decodedPath = URLDecoder.decode(path, "UTF-8");
+        File packageDir = new File(decodedPath);
+
+        // parcourir tous les fichiers dans le répertoire du package
+        File[] files = packageDir.listFiles();
+        if (files != null) {
+            for (File file : files) {
+                if (file.isFile() && file.getName().endsWith(".class")) {
+                    String className = packageName + "." + file.getName().replace(".class", "");
+                    Class<?> classe = Class.forName(className);
+                    if (this.isController(classe)) {
+                        res.add(classe);
+                    }
+                }
+            }
+        }
+        return res;
+
+    }
+}
